IUS-157: Fixed size_t warning for Windows

+++ library/src/iusSourceDict.c
@@ -177,7 +177,7 @@ static int iusSourceDictUpdateKeys
     iusSourceDictDeleteKeys(dict);
     // allocate memory for the keys
     int keyIndex;
-    int size = iusSourceDictGetSize(dict);
+    size_t size = iusSourceDictGetSize(dict);
     dict->keys = calloc(size+1, sizeof(char*));
     IUS_ERR_ALLOC_NULL_N_RETURN(dict, char *, IUS_ERR_VALUE);
